+++ bindings/wasm/examples/merkle_key.js
